+++ subprojects/combined/src/test/groovy/org/opendolphin/core/comm/ClientConnectorPushTests.groovy
@@ -4,7 +4,7 @@ import groovyx.gpars.dataflow.DataflowQueue
 import org.opendolphin.LogConfig
 import org.opendolphin.core.client.ClientDolphin
 import org.opendolphin.core.server.EventBus
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
 import spock.lang.Specification
 
 import java.util.concurrent.CountDownLatch
@@ -13,7 +13,7 @@ import java.util.concurrent.TimeUnit
 class ClientConnectorPushTests extends Specification {
 
     volatile TestInMemoryConfig app
-    ServerDolphin serverDolphin
+    GServerDolphin serverDolphin
     ClientDolphin clientDolphin
 
     protected void setup() {
