+++ subprojects/combined/src/test/groovy/org/opendolphin/core/comm/ControllingWorkflowTests.groovy
@@ -18,7 +18,7 @@ package org.opendolphin.core.comm
 
 import org.opendolphin.LogConfig
 import org.opendolphin.core.client.ClientDolphin
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
 import java.util.concurrent.TimeUnit
 
 /**
@@ -29,7 +29,7 @@ import java.util.concurrent.TimeUnit
 class ControllingWorkflowTests extends GroovyTestCase {
 
     volatile TestInMemoryConfig context
-    ServerDolphin serverDolphin
+    GServerDolphin serverDolphin
     ClientDolphin clientDolphin
 
     @Override
