+++ subprojects/combined/src/test/groovy/org/opendolphin/core/comm/ServerControlledFunctionalTests.groovy
@@ -16,7 +16,6 @@
 
 package org.opendolphin.core.comm
 
-import org.opendolphin.LogConfig
 import org.opendolphin.core.client.ClientAttribute
 import org.opendolphin.core.client.ClientDolphin
 import org.opendolphin.core.server.*
@@ -33,7 +32,7 @@ import java.util.concurrent.TimeUnit
 class ServerControlledFunctionalTests extends GroovyTestCase {
 
     volatile TestInMemoryConfig context
-    ServerDolphin serverDolphin
+    GServerDolphin serverDolphin
     ClientDolphin clientDolphin
 
     @Override
