+++ subprojects/demo-javafx/combined/src/main/groovy/org/opendolphin/demo/team/TeamInMemoryStarter.java
@@ -2,17 +2,11 @@ package org.opendolphin.demo.team;
 
 import groovyx.gpars.agent.Agent;
 import javafx.application.Application;
-import org.opendolphin.core.NoModelStore;
 import org.opendolphin.core.client.ClientDolphin;
-import org.opendolphin.core.client.ClientModelStore;
-import org.opendolphin.core.client.comm.BlindCommandBatcher;
-import org.opendolphin.core.client.comm.HttpClientConnector;
 import org.opendolphin.core.client.comm.InMemoryClientConnector;
-import org.opendolphin.core.client.comm.JavaFXUiThreadHandler;
-import org.opendolphin.core.comm.JsonCodec;
 import org.opendolphin.core.server.DTO;
 import org.opendolphin.core.server.EventBus;
-import org.opendolphin.core.server.ServerDolphin;
+import org.opendolphin.core.server.GServerDolphin;
 import org.opendolphin.demo.JavaFxInMemoryConfig;
 
 import java.util.LinkedList;
@@ -27,7 +21,7 @@ public class TeamInMemoryStarter {
         final JavaFxInMemoryConfig config = new JavaFxInMemoryConfig();
         ClientDolphin clientDolphin = config.getClientDolphin();
         ((InMemoryClientConnector)clientDolphin.getClientConnector()).setSleepMillis(0);
-        ServerDolphin serverDolphin = config.getServerDolphin();
+        GServerDolphin serverDolphin = config.getServerDolphin();
         serverDolphin.register(new TeamMemberActions(teamBus, history));
         serverDolphin.getServerConnector().register(new TeamBusRelease(teamBus));
