+++ subprojects/demo-javafx/server/src/main/groovy/org/opendolphin/demo/PullTreeActionHandler.groovy
@@ -19,7 +19,7 @@ package org.opendolphin.demo
 import org.opendolphin.core.comm.Command
 import org.opendolphin.core.comm.NamedCommand
 import org.opendolphin.core.server.DTO
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
 import org.opendolphin.core.server.Slot
 import org.opendolphin.core.server.comm.NamedCommandHandler
 
@@ -44,7 +44,7 @@ class PullTreeActionHandler implements NamedCommandHandler {
 
     private void makePM(node, List<Command> response) {
         DTO model = new DTO(new Slot("parent", node.parent()?.name()))
-        ServerDolphin.presentationModel(response, node.name(), null, model)
+        GServerDolphin.presentationModel(response, node.name(), null, model)
         node.children().each {
             makePM it, response // recurse
         }
