+++ subprojects/server/src/main/groovy/org/opendolphin/core/server/ServerDolphinFactory.java
@@ -1,0 +1,14 @@
+package org.opendolphin.core.server;
+
+/**
+ * Created by hendrikebbers on 20.01.15.
+ */
+public class ServerDolphinFactory {
+
+    private ServerDolphinFactory() {}
+
+    public static ServerDolphin create() {
+        return new GServerDolphin();
+    }
+
+}
