+++ subprojects/server/src/main/groovy/org/opendolphin/core/server/action/CreatePresentationModelAction.groovy
@@ -21,7 +21,7 @@ import org.opendolphin.core.PresentationModel
 import org.opendolphin.core.Tag
 import org.opendolphin.core.comm.CreatePresentationModelCommand
 import org.opendolphin.core.server.ServerAttribute
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
 import org.opendolphin.core.server.ServerPresentationModel
 import org.opendolphin.core.server.comm.ActionRegistry
 import groovy.transform.CompileStatic
@@ -36,7 +36,7 @@ class CreatePresentationModelAction extends DolphinServerAction {
         }
     }
 
-    private static void createPresentationModel(CreatePresentationModelCommand command, ServerDolphin serverDolphin) {
+    private static void createPresentationModel(CreatePresentationModelCommand command, GServerDolphin serverDolphin) {
         if(serverDolphin.getAt(command.pmId) != null) {
             log.info("Ignoring create PM '$command.pmId' since it is already in the model store.")
             return
