+++ subprojects/server/src/main/groovy/org/opendolphin/server/adapter/DolphinServlet.groovy
@@ -16,10 +16,9 @@
 package org.opendolphin.server.adapter
 
 import groovy.util.logging.Log
-import org.opendolphin.core.ModelStore
 import org.opendolphin.core.comm.Codec
 import org.opendolphin.core.comm.JsonCodec;
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
 import org.opendolphin.core.server.ServerConnector
 import org.opendolphin.core.server.ServerModelStore;
 
@@ -49,13 +48,13 @@ abstract class DolphinServlet extends HttpServlet {
         resp.outputStream.close()
     }
 
-    private ServerDolphin checkDolphinInSession(HttpServletRequest request) {
+    private GServerDolphin checkDolphinInSession(HttpServletRequest request) {
         def session = request.session
-        ServerDolphin dolphin = session.getAttribute(DOLPHIN_ATTRIBUTE_ID)
+        GServerDolphin dolphin = session.getAttribute(DOLPHIN_ATTRIBUTE_ID)
         if (!dolphin) {
             log.info "creating new dolphin for session $session.id"
             def modelStore = new ServerModelStore()
-            dolphin = new ServerDolphin(modelStore, new ServerConnector(codec: codec, serverModelStore: modelStore))
+            dolphin = new GServerDolphin(modelStore, new ServerConnector(codec: codec, serverModelStore: modelStore))
             dolphin.registerDefaultActions()
             registerApplicationActions(dolphin)
             session.setAttribute(DOLPHIN_ATTRIBUTE_ID, dolphin)
@@ -67,5 +66,5 @@ abstract class DolphinServlet extends HttpServlet {
         new JsonCodec()
     }
 
-    protected abstract void registerApplicationActions(ServerDolphin serverDolphin)
+    protected abstract void registerApplicationActions(GServerDolphin serverDolphin)
 }
