+++ subprojects/server/src/test/groovy/org/opendolphin/core/server/action/StoreAttributeActionTests.groovy
@@ -3,17 +3,18 @@ package org.opendolphin.core.server.action
 import org.opendolphin.core.comm.AttributeCreatedNotification
 import org.opendolphin.core.comm.ChangeAttributeMetadataCommand
 import org.opendolphin.core.server.ServerAttribute
-import org.opendolphin.core.server.ServerDolphin
+import org.opendolphin.core.server.GServerDolphin
+import org.opendolphin.core.server.ServerDolphinFactory
 import org.opendolphin.core.server.ServerPresentationModel
 import org.opendolphin.core.server.comm.ActionRegistry
 
 class StoreAttributeActionTests extends GroovyTestCase {
-    ServerDolphin dolphin
+    GServerDolphin dolphin
     ActionRegistry registry
 
     @Override
     protected void setUp() throws Exception {
-        dolphin = new ServerDolphin()
+        dolphin = ServerDolphinFactory.create()
         dolphin.serverModelStore.currentResponse = []
         registry = new ActionRegistry()
     }
