+++ subprojects/server/src/test/groovy/org/opendolphin/server/adapter/DolphinServletSpec.groovy
@@ -1,8 +1,8 @@
 package org.opendolphin.server.adapter
 
 import org.opendolphin.core.comm.Codec
-import org.opendolphin.core.server.ServerDolphin
 import org.opendolphin.core.server.ServerConnector
+import org.opendolphin.core.server.GServerDolphin
 import spock.lang.Specification
 
 import javax.servlet.ServletInputStream
@@ -42,7 +42,7 @@ class DolphinServletSpec extends Specification {
     void "calling doPost in existing session must not reach registration of custom actions"() {
         given:
         DolphinServlet.log.level = Level.ALL // for full branch coverage
-        def (servlet, req, resp) = mockServlet(new ServerDolphin(null, mockServerConnector()))
+        def (servlet, req, resp) = mockServlet(new GServerDolphin(null, mockServerConnector()))
 
         when:
         servlet.doPost(req, resp)
@@ -60,8 +60,8 @@ class DolphinServletSpec extends Specification {
         ] as ServerConnector
     }
 
-    def mockServlet(ServerDolphin serverDolphin) {
-        ServerDolphin.metaClass.getServerConnector = {-> mockServerConnector() }
+    def mockServlet(GServerDolphin serverDolphin) {
+        GServerDolphin.metaClass.getServerConnector = {-> mockServerConnector() }
 
         def servlet = new TestDolphinServlet()
         def session = [
@@ -85,7 +85,7 @@ class TestDolphinServlet extends DolphinServlet {
     boolean registerReached
 
     @Override
-    protected void registerApplicationActions(ServerDolphin serverDolphin) {
+    protected void registerApplicationActions(GServerDolphin serverDolphin) {
         registerReached = true
     }
