+++ subprojects/shared/src/main/groovy/org/opendolphin/core/ModelStore.java
@@ -27,23 +27,23 @@ import java.util.*;
  * both on the client (view) and on the server (controller) side for separate access.
  */
 
-public class ModelStore {
+public class ModelStore<U extends Attribute, T extends PresentationModel<U>> {
 
     // We maintain four indexes in this data structure in order to efficiently access
     // - presentation models by id or by type
     // - attributes by id or by qualifier
 
-    private final Map<String, PresentationModel>        presentationModels;
-    private final Map<String, List<PresentationModel>>  modelsPerType;
-    private final Map<String, Attribute>                attributesPerId;
-    private final Map<String, List<Attribute>>          attributesPerQualifier;
+    private final Map<String, T>        presentationModels;
+    private final Map<String, List<T>>  modelsPerType;
+    private final Map<String, U>                attributesPerId;
+    private final Map<String, List<U>>          attributesPerQualifier;
 
     private final Set<ModelStoreListenerWrapper> modelStoreListeners = new LinkedHashSet<ModelStoreListenerWrapper>();
 
     private final PropertyChangeListener ATTRIBUTE_WORKER = new PropertyChangeListener() {
         @Override
         public void propertyChange(PropertyChangeEvent event) {
-            Attribute attribute = (Attribute) event.getSource();
+            U attribute = (U) event.getSource();
             String oldQualifier = (String) event.getOldValue();
             String newQualifier = (String) event.getNewValue();
 
@@ -58,10 +58,10 @@ public class ModelStore {
     }
 
     public ModelStore(ModelStoreConfig config) {
-        presentationModels      = new HashMap<String, PresentationModel>        (config.getPmCapacity());
-        modelsPerType           = new HashMap<String, List<PresentationModel>>  (config.getTypeCapacity());
-        attributesPerId         = new HashMap<String, Attribute>                (config.getAttributeCapacity());
-        attributesPerQualifier  = new HashMap<String, List<Attribute>>          (config.getQualifierCapacity());
+        presentationModels      = new HashMap<String, T>        (config.getPmCapacity());
+        modelsPerType           = new HashMap<String, List<T>>  (config.getTypeCapacity());
+        attributesPerId         = new HashMap<String, U>                (config.getAttributeCapacity());
+        attributesPerQualifier  = new HashMap<String, List<U>>          (config.getQualifierCapacity());
     }
 
     /**
@@ -80,7 +80,7 @@ public class ModelStore {
      *
      * @return a {@code Collection} of all presentation models found in this store.
      */
-    public Collection<PresentationModel> listPresentationModels() {
+    public Collection<T> listPresentationModels() {
         return Collections.unmodifiableCollection(presentationModels.values());
     }
 
@@ -92,7 +92,7 @@ public class ModelStore {
      * @param model the model to be added.
      * @return if the add operation was successful or not.
      */
-    public boolean add(PresentationModel model) {
+    public boolean add(T model) {
         if (null == model) return false;
 
         if (presentationModels.containsKey(model.getId())) {
@@ -102,7 +102,7 @@ public class ModelStore {
         if (!presentationModels.containsValue(model)) {
             presentationModels.put(model.getId(), model);
             addPresentationModelByType(model);
-            for (Attribute attribute : model.getAttributes()) {
+            for (U attribute : model.getAttributes()) {
                 addAttributeById(attribute);
                 attribute.addPropertyChangeListener(Attribute.QUALIFIER_PROPERTY, ATTRIBUTE_WORKER);
                 if (!StringUtil.isBlank(attribute.getQualifier())) addAttributeByQualifier(attribute);
@@ -119,13 +119,13 @@ public class ModelStore {
      * @param model the model to be removed from the store.
      * @return if the remove operation was successful or not.
      */
-    public boolean remove(PresentationModel model) {
+    public boolean remove(T model) {
         if (null == model) return false;
         boolean removed = false;
         if (presentationModels.containsValue(model)) {
             removePresentationModelByType(model);
             presentationModels.remove(model.getId());
-            for (Attribute attribute : model.getAttributes()) {
+            for (U attribute : model.getAttributes()) {
                 removeAttributeById(attribute);
                 removeAttributeByQualifier(attribute);
                 attribute.removePropertyChangeListener(Attribute.QUALIFIER_PROPERTY, ATTRIBUTE_WORKER);
@@ -136,55 +136,55 @@ public class ModelStore {
         return removed;
     }
 
-    protected void addAttributeById(Attribute attribute) {
+    protected void addAttributeById(U attribute) {
         if (null == attribute || attributesPerId.containsKey(attribute.getId())) return;
         attributesPerId.put(attribute.getId(), attribute);
     }
 
-    protected void removeAttributeById(Attribute attribute) {
+    protected void removeAttributeById(U attribute) {
         if (null == attribute) return;
         attributesPerId.remove(attribute.getId());
     }
 
-    protected void addAttributeByQualifier(Attribute attribute) {
+    protected void addAttributeByQualifier(U attribute) {
         if (null == attribute) return;
         String qualifier = attribute.getQualifier();
         if (StringUtil.isBlank(qualifier)) return;
-        List<Attribute> list = attributesPerQualifier.get(qualifier);
+        List<U> list = attributesPerQualifier.get(qualifier);
         if (null == list) {
-            list = new ArrayList<Attribute>();
+            list = new ArrayList<U>();
             attributesPerQualifier.put(qualifier, list);
         }
         if (!list.contains(attribute)) list.add(attribute);
     }
 
-    protected void removeAttributeByQualifier(Attribute attribute) {
+    protected void removeAttributeByQualifier(U attribute) {
         if (null == attribute) return;
         String qualifier = attribute.getQualifier();
         if (StringUtil.isBlank(qualifier)) return;
-        List<Attribute> list = attributesPerQualifier.get(qualifier);
+        List<U> list = attributesPerQualifier.get(qualifier);
         if (null != list) {
             list.remove(attribute);
         }
     }
 
-    protected void addPresentationModelByType(PresentationModel model) {
+    protected void addPresentationModelByType(T model) {
         if (null == model) return;
         String type = model.getPresentationModelType();
         if (StringUtil.isBlank(type)) return;
-        List<PresentationModel> list = modelsPerType.get(type);
+        List<T> list = modelsPerType.get(type);
         if (null == list) {
-            list = new ArrayList<PresentationModel>();
+            list = new ArrayList<T>();
             modelsPerType.put(type, list);
         }
         if (!list.contains(model)) list.add(model);
     }
 
-    protected void removePresentationModelByType(PresentationModel model) {
+    protected void removePresentationModelByType(T model) {
         if (null == model) return;
         String type = model.getPresentationModelType();
         if (StringUtil.isBlank(type)) return;
-        List<PresentationModel> list = modelsPerType.get(type);
+        List<T> list = modelsPerType.get(type);
         if (null == list) return;
         list.remove(model);
         if (list.isEmpty()) {
@@ -192,9 +192,9 @@ public class ModelStore {
         }
     }
 
-    protected void removeAttributeByQualifier(Attribute attribute, String qualifier) {
+    protected void removeAttributeByQualifier(U attribute, String qualifier) {
         if (StringUtil.isBlank(qualifier)) return;
-        List<Attribute> list = attributesPerQualifier.get(qualifier);
+        List<U> list = attributesPerQualifier.get(qualifier);
         if (null == list) return;
         list.remove(attribute);
         if (list.isEmpty()) {
@@ -209,7 +209,7 @@ public class ModelStore {
      * @param id the id to search
      * @return a presentation model instance of there's an id match, {@code null} otherwise.
      */
-    public PresentationModel findPresentationModelById(String id) {
+    public T findPresentationModelById(String id) {
         return presentationModels.get(id);
     }
 
@@ -220,7 +220,7 @@ public class ModelStore {
      * @param type the type to search for
      * @return a {@code List} of all presentation models for which there was a match in their type.
      */
-    public List<PresentationModel> findAllPresentationModelsByType(String type) {
+    public List<T> findAllPresentationModelsByType(String type) {
         if (StringUtil.isBlank(type) || !modelsPerType.containsKey(type)) return Collections.emptyList();
         return Collections.unmodifiableList(modelsPerType.get(type));
     }
@@ -242,7 +242,7 @@ public class ModelStore {
      * @param id the id to search for.
      * @return an attribute whose id matches the parameter, {@code null} otherwise.
      */
-    public Attribute findAttributeById(String id) {
+    public U findAttributeById(String id) {
         return attributesPerId.get(id);
     }
 
@@ -252,12 +252,12 @@ public class ModelStore {
      *
      * @return a {@code List} of all attributes fo which their qualifier was a match.
      */
-    public List<Attribute> findAllAttributesByQualifier(String qualifier) {
+    public List<U> findAllAttributesByQualifier(String qualifier) {
         if (StringUtil.isBlank(qualifier) || !attributesPerQualifier.containsKey(qualifier)) return Collections.emptyList();
         return Collections.unmodifiableList(attributesPerQualifier.get(qualifier));
     }
 
-    public void registerAttribute(Attribute attribute) {
+    public void registerAttribute(U attribute) {
         if (null == attribute) return;
         boolean listeningAlready = false;
         for (PropertyChangeListener listener : attribute.getPropertyChangeListeners(Attribute.QUALIFIER_PROPERTY)) {
@@ -303,7 +303,7 @@ public class ModelStore {
                 modelStoreListeners.contains(new ModelStoreListenerWrapper(presentationModelType, listener));
     }
 
-    protected void fireModelStoreChangedEvent(PresentationModel model, ModelStoreEvent.Type eventType) {
+    protected void fireModelStoreChangedEvent(T model, ModelStoreEvent.Type eventType) {
         if (modelStoreListeners.isEmpty()) return;
         ModelStoreEvent event = new ModelStoreEvent(eventType, model);
         for (ModelStoreListener listener : modelStoreListeners) {
